Quiet initial release for brave users.
Use at your own risk. There is a lot to do, but it makes sense to at
least get this code out there. Large architectural changes will probably
land in the future, in particular with cobbler v2.4+ This was tested
with cobbler v.2.2.3 from epel.

If it's not clear, I'm releasing this because it's useful to me, but maybe
it's useful to you if you're in a similar position as i'm in. Cheers!